agents/dars-agent/requirements.txt
anthropic[bedrock]
anthropic[vertex]
config
datasets
docker >= 7.1.0
gymnasium
numpy
openai>=1.0
pandas
rich
ruamel.yaml
swebench==2.1.1
tenacity
unidiff
simple-parsing
together>=1.1.0
ollama
rich-argparse
flask
flask-cors
flask-socketio
voyageai
faiss-cpu
networkx
google-generativeai
langchain-voyageai
litellm
grep_ast
hydra-core